fixed the pr comments.
