+++ app/src/main/java/com/clint/nasa/core/Constants.kt
@@ -1,0 +1,5 @@
+package com.clint.nasa.core
+
+object Constants {
+    const val SPLASH_SCREEN_DELAY =  2000L
+}
