+++ app/src/main/java/com/clint/nasa/pictures/PicturesRepository.kt
@@ -2,8 +2,8 @@ package com.clint.nasa.pictures
 
 import com.clint.nasa.core.exception.Failure
 import com.clint.nasa.core.functional.Either
+import com.clint.nasa.core.functional.Request.request
 import com.clint.nasa.core.platform.NetworkHandler
-import retrofit2.Call
 import javax.inject.Inject
 
 interface PicturesRepository {
@@ -13,8 +13,7 @@ interface PicturesRepository {
 
 class NetWork
 @Inject constructor(
-    private val networkHandler: NetworkHandler,
-    private val pictureService: PictureService
+    private val networkHandler: NetworkHandler, private val pictureService: PictureService
 ) : PicturesRepository {
 
     override fun pictures(): Either<Failure, List<Pictures>> {
@@ -23,8 +22,7 @@ class NetWork
                 request(
                     pictureService.getPictures(), {
                         it
-                    },
-                    mutableListOf()
+                    }, mutableListOf()
                 )
             }
             false -> {
@@ -35,20 +33,3 @@ class NetWork
 
 }
 
-private fun <T, R> request(
-    call: Call<T>,
-    transform: (T) -> R,
-    default: T
-): Either<Failure, R> {
-    return try {
-        val response = call.execute()
-        when (response.isSuccessful) {
-            true -> {
-                Either.Right(transform((response.body() ?: default)))
-            }
-            false -> Either.Left(Failure.ServerError)
-        }
-    } catch (exception: Throwable) {
-        Either.Left(Failure.ServerError)
-    }
-}
